DO-NOT-MERGE: this needs some repairs first

+++ virtdata-lang/src/main/java/io/nosqlbench/virtdata/lang/ast/FloatArg.java
@@ -24,7 +24,7 @@ public class FloatArg implements ArgType {
         this.floatValue = floatValue;
     }
 
-    public double getFloatValue() {
+    public float getFloatValue() {
         return floatValue;
     }
